requirements.txt
streamlit>=1.22.0
playwright>=1.30.0
pandas>=1.5.3
pillow>=9.5.0
pytesseract>=0.3.10
groq>=0.4.0
python-dotenv>=1.0.0